Reorganisation of the config
