+++ backend/src/main/java/com/glassait/tanks_and_equipments/shared/configs/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.glassait.tanks_and_equipments.config;
+package com.glassait.tanks_and_equipments.shared.configs;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
